refactor(storage-resize-images): move files about
